Add arbirtary UID support
This will allow users to run the image with any arbitrary UID.  This helps run
the image in a more secure k8s environments that have locked down
securityContexts such as runAsNonRoot.  This also allows for flexability when
fsGroup is set and this is being used as a sidecar for debugging.